Changes, separate list and listbox
