+++ packages/list/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./list"
+export * from "./list-box"
