+++ packages/typography/src/blockquote.tsx
@@ -115,10 +115,6 @@ export const Blockquote: React.FC<BlockquoteProps> = (props) => {
       bgGradientMiddleColorContrast: props.bgCaptionGradientMiddleColorContrast,
       bgGradientEndColor: props.bgCaptionGradientEndColor,
       bgGradientEndColorContrast: props.bgCaptionGradientEndColorContrast,
-      borderWidth: props.borderWidth,
-      borderColor: props.borderColor,
-      borderColorContrast: props.borderColorContrast,
-      borderStyle: props.borderStyle,
       borderRadius: props.rounded,
       borderRadiusPosition: (props.rounded !== undefined && props.caption !== undefined) ? "bottom" : props.roundedPosition,
       shadow: props.shadow
@@ -149,7 +145,9 @@ export const Blockquote: React.FC<BlockquoteProps> = (props) => {
         {props.quote && (
           <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" className={`my-3 fill-current ${clsQuote} opacity-20`}><path d="M13 14.725c0-5.141 3.892-10.519 10-11.725l.984 2.126c-2.215.835-4.163 3.742-4.38 5.746 2.491.392 4.396 2.547 4.396 5.149 0 3.182-2.584 4.979-5.199 4.979-3.015 0-5.801-2.305-5.801-6.275zm-13 0c0-5.141 3.892-10.519 10-11.725l.984 2.126c-2.215.835-4.163 3.742-4.38 5.746 2.491.392 4.396 2.547 4.396 5.149 0 3.182-2.584 4.979-5.199 4.979-3.015 0-5.801-2.305-5.801-6.275z"/></svg>
         )}
-        <p className={clsText}>{props.children}</p>
+        {typeof props.children === "string" ? (
+          <p className={clsText}>{props.children}</p>
+        ) : (props.children)}
       </blockquote>
       {props.caption !== undefined && (
         <figcaption className={clsCaption}>{props.caption}</figcaption>
@@ -160,10 +158,15 @@ export const Blockquote: React.FC<BlockquoteProps> = (props) => {
 
 Blockquote.defaultProps = {
   quote: true,
-  bgColor: "gray",
-  bgColorContrast: "200",
+  bgColor: "white",
+  border: true,
+  borderWidth: "normal",
+  borderStyle: "solid",
+  borderColor: "gray",
+  borderColorContrast: "200",
   rounded: "md",
-  py: "2",
+  fontSize: "base",
+  py: "4",
   px: "3",
   mx: "3",
   my: "3"
